Availability / Last Message Time
Appends a timestamp indicating the time the last message was broadcast from this device. 

I have a terminal becomes unresponsive after a random period of time and needs rebooting. However, the device stats in HA persist and so the device appears as if it is online. Broadcast of a timestamp with each sensor update makes it possible to see when the device last communicated with MQTT in Home Assistant.